Session/tab sync performance tests.
This patch adds tab sync performance tests, similar to the tests for the existing datatypes in live_sync/performance.  Changes to LiveSessionsSyncTest were required - most significantly the addition of AwaitSessionPropagation, which waits until a vector of urls can all be found in the model associator.  This type of wait loop previously existed as part of OpenTab, and worked with a single url.  AwaitSessionPropagation supports multiple urls, and is called from OpenTab, OpenMultipleTabs, and SessionsSyncPerfTest::UpdateTabs.

BUG=none
TEST=sync_performance_tests, sync_integration_tests

Review URL: http://codereview.chromium.org/7575026