Fix some egregious bugs in Var.
Self-assignment was broken and would lose the reference. I uncovered this when
running a test. It outputted a warning to the console, but we never looked at
it. I made a more explicit test.

This also fixes output exceptions. The OutException helper class detected
whether the existing object had an exception or not incorrectly. This was
exposed when var assignment was fixed.

TEST=included
BUG=none
Review URL: http://codereview.chromium.org/7511026