Use MessageLoopProxy instead of MessageLoop to dispatch notifications in ObserverListThreadsafe.
This is nearly identical to the reverted r96013, with the modification that
std::map::insert isn't used.  Its invocation was complex (such that
it worked in some compilers but not in others) and the performance gain is
negligible in any use case I've seen in Chromium.

BUG=91589

Review URL: http://codereview.chromium.org/7604006