chromeos: Try to avoid LoginUtilsImpl delegate segfault.
This makes LoginUtilsImpl::OnProfileCreated() check that its
delegate is non-NULL before using it.  I'm hoping that this
will avoid a segfault that's apparently showing up during
testing when Chrome is killed during initialization.

Note that this is just a temporary workaround; the
underlying code is possibly still incorrect (or relying on
incorrect assumptions about the order in which things will
happen).

BUG=chromium-os:18269
TEST=none
TBR=altimofeev@chromium.org

Review URL: http://codereview.chromium.org/7590026