views-desktop: Fix sending mousewheel events to the correct view.
Mousewheel events, like keyboard events and unlike other mouseevents, are sent
to the focused view from the focus-manager, instead of sending to the view at
the location of the LocatedEvent. So This needs to be handled from the
dekstop/window-manager to send the events to the focused window.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7462017