Allow pyauto performance tests in perf.py to run as an autotest.
This change adds autotest control files and an associated pyauto suite
for running the tests in perf.py as an autotest.  This change only
allows execution as an autotest, but does not enable the tests to run
continuously in the autotest lab.  That will be done in a later CL.

I also modified perf.py so that the v8 benchmark test only runs for a
single iteration, and for the other tests, the number of iterations
can optionally be tweaked by setting an environment variable (useful when
running the perf tests manually).

BUG=chromium-os:18187,chromium-os:18185,chromium-os:18458
TEST=None

Review URL: http://codereview.chromium.org/7617014