Update placement of interface 'Define'
A minor change to the way we output the headers.  This change moves the
defines which are used to name the interfaces and thier versions to the
top of the header, moving out of the @file block.  This prevents them
from being 'spewed' in the documentation.

BUG= none
TEST= none
Review URL: http://codereview.chromium.org/7669001