Revert 97246 - Merge 96904 - Add option to not generate resources on bind in OpenGL ES
This allowes us to more efficiently manage ids. It is
not OpenGL ES 2.0 compatible though it probably fits most OpenGL ES
programs.

Note that we need to turn this off on Pepper and/or probably
provide a way for Pepper to turn on on. I'm not sure of the
path Pepper takes to setup. Assuming it goes through
GraphicsContext3D then changes to webkit will be needed to
get the flag all the way down through IPC to the GPU process.

TEST=unit tests and ran a few pages in a chrome build
BUG=92260

Review URL: http://codereview.chromium.org/7633060

TBR=gman@chromium.org
Review URL: http://codereview.chromium.org/7655036

TBR=apatrick@chromium.org
Review URL: http://codereview.chromium.org/7624034