Chrome OS: don't display connection failure notification unless connect was user-initiated.
When a network connection attempt is initiated because of auto-connect,
it doesn't make sense to notify the user of a failure. The case in which
connection failure can be expected is across a suspend/resume cycle,
e.g., when the netbook lid is closed and reopened.

BUG=chromium-os:16391
TEST=On a chromebook with an active cellular account, make sure that
the "Automatically connect to this network" box is checked. Establish
a connection on the cellular network. Close the chromebook lid. Wait
about 20 seconds, and then re-open the lid. After 10-15 seconds, the
"Disable mobile data" button should reappear on the Settings > Internet
page, and then a while later, a connection on the cellular network
should be established, with no notification pop-up. The original
problem seems more repeatable if Wi-Fi is disabled.

R=stevenjb@chromium.org

Review URL: http://codereview.chromium.org/7631045