Fix WaitForSingleNonChildProcess. There were three bugs in the loop usage.
There was a bug in the original code in that if the loop executes a second
time, kq will have been closed.

The existing code was not careful about resetting the timeout in the event
that the kevent had to be restarted in the loop.

The existing code would continue adding the same "change" to the kqueue on
each iteration of the loop.

The replacement code has been taken mostly from
content/common/process_watcher_mac.cc WaitForChildToDie. I would actually like
to align these two functions to use the same backend implementation, but that
will require better signalling from the backend about which specific case is
causing it to return, because WaitForChildToDie needs to be able to reap or
kill-and-reap the process while avoiding the "zombie kernel death race"
explained in that function's comments. Since WaitForSingleNonChildProcess is
currently only concerned with non-child processes, it doesn't need to worry
about any of this reaping business.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/7708006