Revert 97808 - Make WeakPtr thread-safe, i.e. allow cross-thread copying of WeakPtr and destruction on a thread other than the one where the original reference was created.
The problem with the current implementation is modifying the flag pointer
from WeakPtr::~WeakPtr which might happen on a different thread (potentially
racing with somebody invalidating the flag on the original thread).

For compatibility reasons, creating the initial reference attaches to the
thread and governs the thread-safety checking logic with respect to checking
a reference to be valid and invalidating it, which should both only be done
on the same thread.

BUG=82509
TEST=added unit tests

Review URL: http://codereview.chromium.org/7677028

TBR=sievers@chromium.org
Review URL: http://codereview.chromium.org/7685054