Measure sync unit test coverage
Adds the sync_unit_test binary to the list used to generate
http://build.chromium.org/f/chromium/coverage/

This should be fairly safe, since it doesn't touch the important parts of the build files.  However, I am a bit worried that I haven't tested it on a build-bot.  The gyp-generated build files seem to behave as expected, though.

BUG=
TEST=

Review URL: http://codereview.chromium.org/7694005