Remove IdFromNativeView, move this logic to RenderWidgetHostViwe::GetNativeViewId
[Linux/Mac] Changed RenderWidgetHost::GetScreenInfo/GetWindowRect/GetRootWindowRect to use RWHV instead of NativeViewId as use of NativeViewId is no longer necessary.

This is necessary to allow us to switch RWHV implementations
(RenderWidgetHostViewGtk and RenderWidgetHostViewViews) on the fly. (Windows will need something similar to support NPAPI with pure views, but this part is not implemented)
The current code doesn't work with RenderWidgetHostWin as win uses IPC method to get screen/window information and needs more work.

BUG=none
TEST=none No functional change. All test should pass.

Review URL: http://codereview.chromium.org/7670039