Call GetPluginsInfo() to register flash plugin before attempting to load on Mac 10.5
It seems to take some time before Flash files can be loaded on Leopard.
These tests open a flash file right after launching the browser, and instead
of opening the flash, the browser downloads it, so it seems like the plugin
info isn't registered yet. This used to work in the past, so it's a
regression. Also, this happens only on 10.5

This CL works around the problem by calling GetPluginsInfo() which would
trigger the plugins to get registered.

BUG=94123

R=dennisjeffrey@chromium.org
TEST=

Review URL: http://codereview.chromium.org/7727002