CL removing inheritance of SkDevice from PlatformDevice.
PlatformDevice is now a base interface, which is implemented by the various flavours of BitmapPlatformDevice, and VectorPlatformDevice.  The BitmapPlatformDevice and VectorPlatformDevice classes now inherit directly from SkDevice, or SkPDFDevice, as appropriate.

PlatformDevice helper functions access the PlatformDevice interface attached to a SkDevice via meta-data on the SkDevice.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/7633040