Lock the initialization of the GL bindings.
The GpuInfoCollector and CompositorGL exist on different threads in the browser process and both use the GL bindings. I've witnessed a race in the binding initialization where one thread thinks that the bindings have been initialized, but the initialization hasn't completed on the other thread, and a NULL function pointer gets dereferenced.

BUG=none
TEST=fixes segfault on TOUCH_UI build

Review URL: http://codereview.chromium.org/7724003