Make a channel and time limited persistent PreRead experiment
Our previous PreRead experiment involved flipping a coin on every launch, but we wish to persist the coin-toss between launches. Additionally, this version of the field trial includes an expiry date and limits it to the canary and dev channels.

We are unable to use the FieldTrial mechanism due to PreRead occuring prior to chrome.dll being loaded. As such, we use the registry for persisting the coin-toss and environment variables for communicating between chrome.exe and chrome.dll.

Review URL: http://codereview.chromium.org/7712014