Revert 98392 - Split WebGraphicsContext3DCommandBufferImpl::initialize() into two stages.
After the context is created, the remaining calls should be done on
the compositor thread. This CL adds a MaybeInitializeGL() method, which
we run on each call to makeContextCurrent().

Added asserts to RendererGLContext to check that it's always used on
the same thread.

Review URL: http://codereview.chromium.org/7713015

TBR=husky@google.com
Review URL: http://codereview.chromium.org/7745042