Block {am, fa, sw} and  {kn, te} on CrOS and roll ICU to 98343
The former is blocked until we have translation back for them. For ChromeOS R14, we'll not have them ready.

The former is blocked until we have fonts for them to use on CrOS. 
This change will be reverted once two issues above are resolved.

See http://codereview.chromium.org/7725002/ for the ICU part that
removed the CrOS-specific ICU data file.

BUG=chromium-os:19465
TEST=The UI language list does not include the above 5 languages on CrOS.
Review URL: http://codereview.chromium.org/7726003