Fix localized "platform" message for non-8bit languages
Now we use localized string for "platform", we should not
assume the text is encoded in ASCII.

To test this, manually change the UI to a non-latin language
(e.g. Japanese) and see if any irregular font is not displayed.

Webui login screen is also affected.

BUG=chromium-os:19822
TEST=manually

Review URL: http://codereview.chromium.org/7791020