Make sure to re-index the page after a redirect.
The code used to assume that reindexing was only needed if the page id had changed, but that doesn't work in the case of a redirect.  If the page id is unchanged, we'll now check the toplevel URL and reindex if it has changed.

BUG=80742
TEST=none


Review URL: http://codereview.chromium.org/7744042