[Mac] Better fix for MenuController ZombieObjectCrash.
Use a |chrome_application_mac::ScopedSendingEvent| to block window close events for the duration of the context menu.

In addition, instead of retaining the controller for the duration of the menu being open, close the menu and unset the menu's delegate when the controller is destroyed.

BUG=90412
TEST=Using test.close.html from the bug report, open a context menu and wait for the tab to close. When it does, the context menu should be closed and no crash should happen. Also, monitor crash reports to make sure these crashes disappear.

Review URL: http://codereview.chromium.org/7812026