Add some more temporary instrumentation for diagnosing bug 93747.
The goal is to conclusively tell whether the object is freed or not when running TabStripModel::GetWrapperIndex. Also surrounded the vector that we are seeing as corrupt with some special padding bytes (64 bytes total), to see if that area of memory is getting clobbered.

BUG=93747

Review URL: http://codereview.chromium.org/7827024