Let `gclient runhooks` pull clang on mac.
If nothing needs to be done, the script takes 0s to run. If a new
version is available, it's downloaded from the web. It's a 13MB
download, which takes ~20s. This has to be paid ~once/week by all
mac devs, but since clang compiles faster than gcc this is actually
a big speed-up.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/7835005