Don't crash if DontProceed is chosen on the SSL interstitial page after Proceed is chosen.
A common place where this happens is when the user navigates to an https:// URL with a
bad certificate, and the content is a download. The user chooses to proceed, but the SSL interstitial
is never hidden because there is never a page to be committed. The user presses the back button, which
does a DontProceed, and this would crash. Note that "back to safety" and "proceed" within the SSL page
are filtered out after an action is taken, so these are no-ops currently.

This sequence will still treat the SSL certificate as accepted for future navigations.

BUG=90864,93605
TEST=Start run_testserver --https-cert=expired. Navigate to a normal page like www.google.com. Navigate to the testserver at https://localhost:<port>/files/downloads/dangerous/dangerous.exe. See the interstitial and click proceed. Press the back button on the browser, and you should see the normal page again, instead of a crash.


Review URL: http://codereview.chromium.org/7826036