[Mac] Tear down extension-installed bubble before main window closes.
The main window closed before the bubble, so when the bubble tried to
remove itself as a child window, it called a stale reference.

Additionally, revise a couple uses of this pattern to be more robust
by asking the window being closed to remove itself from its current
parent window (which would be nil in this case).

BUG=94172
TEST=see bug, monitor crash server.


Review URL: http://codereview.chromium.org/7740060