Merge 100088 - [Sync] Don't attempt to show the configure screen if the backend isn't ready.
We simply return a null flow, which is the same as if we had done nothing. Next
time the user clicks customize, they will again try to create a new flow, and
if the backend has now initialized, we succeed.

BUG=93127
TEST=manually alter the code to ensure the backend doesn't initialize. Attempt
to customize sync on a client that is already signed in.


Review URL: http://codereview.chromium.org/7851003

TBR=zea@chromium.org
Review URL: http://codereview.chromium.org/7831055