Scale BufferedResourceLoader buffering window based on bit rate of video
This CL adds logic in the FFmpegDemuxer to compute a suggested buffering window
size based on the bit rate and playback rate of the video, then propogates the
suggested size to BufferedResourceLoader upon initialization or changes in
playback rate. BufferedResourceLoader resizes the buffer as necessary.

BUG=60215
TEST=test_shell_tests,media_unittests

Review URL: http://codereview.chromium.org/7718017