Prevent the renderer from dying if launched under App-V/SoftGrid.
A version of Microsoft Office launches chrome under a virtual
environment, and this virtual environment does not allow us
to reduce the integrity level on the renderer.  Since it returns
a fake error (ERROR_INVALID_HANDLE), we can trap it and continue
execution anyway... until we find a better fix.

BUG=95888
Review URL: http://codereview.chromium.org/7857022