Add a copy of the SMHasher library.  MurmurHash3 will be used by SafeBrowsing phishing detection.
We'll be modifying the page_words field in ClientSideModel with murmurhash3's of the words, rather than SHA-256 hashes, since SHA-256 is fairly expensive to compute.  Note that on a murmurhash3 hit, we'll still do a SHA-256 check against page_terms, so this will not introduce term collisions.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/7848010