Suppress navigation events for frames until the navigation of the main frame is committed
A subframe that navigates the main frame can still navigate itself. However, the frame is about to be deleted, and there's no way for the extension to know that it doesn't belong to the new main frame, so I decided to just skip those events.

BUG=none
TEST=ExtensionApiTest.WebNavigationIFrame

Review URL: http://codereview.chromium.org/7851016