[Sync] Don't notify the transaction observer of huge mutation sets.
The transaction observer is currently only used for debug information sent to
chrome://sync_internals. When we have large numbers of changes within a single
transaction, our mutation set can grow exceedingly large, and the unnecessary
copying we perform in sending the data to the transaction observer can result
in out of memory crashes. This change limits the size of the mutation sets
we copy to 1000 elements. If the size is larger, we send a dummy empty mutation
set instead.

BUG=90169
TEST=turn on encryption on an account with >1000 items. Ensure none are printed in the events pane of about:sync


Review URL: http://codereview.chromium.org/7861005