extract out removal of the g_log_bug53991 global.
This global is referenced across content and chrome, and would need
to be exported for the content.dll build. It appears to be off by
default and was added for a bug last touched a year ago, so I'm thinking
it's safe to remove it and it can be readded back in as necessary.

R=mpcomplete@chromium.org
BUG=53991,90442
TEST=none

Review URL: http://codereview.chromium.org/7833004