Third try at committing this.
Patchset 1 is the original patch ( http://codereview.chromium.org/7833042/ ).
Patchset 2 is the original patch plus a fix to a memory leak from the initial commit ( http://codereview.chromium.org/7831056/ )

On the first try there was a memory leak in a test. I fixed that, but made a mistake in the commit (the committed code did not correspond to the reviewed code). Both commits were reverted.

I then landed a new CL ( http://codereview.chromium.org/7860039/ ) that contained the correct changes combining the first two CLs. This caused an error in heapchecker for which a suppression has subsequently been defined ( http://codereview.chromium.org/7780010 ).

In summary, all of this is reviewed, minus some lint fixes.

BUG=68657
TEST=net_unittests / DeferredCookieTaskTest.* and CookieMonsterTest.*
TBR=estade
Review URL: http://codereview.chromium.org/7891008