Reland r101259 after fixing shared-library build issues.
Removed the dependency of PepperPluginRegistry on Pepper proxy.

We would need to move PepperPluginRegistry to webkit/plugins/ppapi so that it can be used by test_shell. But before we can move it, we need to remove all dependencies on content and pepper/proxy.
TEST=Manually ran pepper 3d demos in-process and out-of-process.
TBR=brettw@chromium.org
Review URL: http://codereview.chromium.org/7912001