External textures outlive layers.
When the View is hidden, we destroy the associated Layer, but the external texture should persist.

Similarly, if a View is made visible, we create a Layer and  the Layer should have that external texture.

The solution is to push the external texture logic and state up into the LayerHelper, which persists.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7904024