Allow masters to specify buildbot-derived chrome revisions.
This allows the informational bots to do useful work in slow periods,
and to have that work align with what a chrome dev would expect.
It also gets reported correctly for a ChromeOS dev since the version
link inside the steps is also correct.

BUG=chromium-os:20139
TEST=manual runs (triggered and forced)

Review URL: http://codereview.chromium.org/7883004