Force synchrous writing of "ExitedCleanly" prefs during shutdown
This will cause a regression in shutdown performance by
blocking the UI thread when we are writing the prefs
during shutdown.  This will cause a hang if we are unable
to write the prefs.

I've also added a second crashing hang detector to the "fast
shutdown" path.  It is possible that this will not have time
to trigger, as we may be terminated first... but at least
we're covering all bases (that we can).

As a result, if we hang precludes writing the pref, we'll
see the cause in the hang in a chrash dump.

r=rtenneti
BUG=95375
Review URL: http://codereview.chromium.org/7925021