Revert 101474 - Switch on use of precompiled headers for VS 2008.
The precompiled header brings in more definitions of abs() for the base/time_win.cc file, and a hint to the compiler was needed to let VS 2008 build the file correctly.

On VS 2008 only, the mini_installer target did not link correctly with precompiled headers enabled so this change disables them for that project.

BUG=none
TEST=it builds (esp. on Windows), existing tests pass

Review URL: http://codereview.chromium.org/7910002

TBR=joi@chromium.org
Review URL: http://codereview.chromium.org/7945001