Fix crash bug at net::WebSocketJob::OnSentData.
According to crash dump, I suspect it is crashed at
current_buffer_->DidConsume() where current_buffer_ is NULL.

If current_buffer_ is NULL, WebSocketJob doesn't think there are writing data
and nothing to do, so this CL changes to just return from
net::WebSocketJob::OnSentData in this case.
Also checks if amount_sent is negative or zero and return if it is so.

BUG=91591
TEST=none

Review URL: http://codereview.chromium.org/7923004