Rename PPB_Fullscreen_Dev to PPB_FlashFullscreen.
This is to make way for the new PPB_Fullscreen interface which will hopefully replace the old one at some point. This maintains backwards binary compat.

I renamed two things related to the broker because the file wasn't being included in the proxy properly, and we never noticed they were wrong.

This also fixes a crash in the test harness generating the list of tests, since there is no current test case for the DidChangeView call.
Review URL: http://codereview.chromium.org/7917019