Aura: Use extension-based file picker from CrOS for pure-views builds.
Removes the dependency on chromeos-specific code in the extension-based file picker.  Compiles and links, but does not yet run - the extension subsystem is not finding the resources for the file_manager component extension.  The Aura build currently won't process keyboard or mouse events, so I don't have a good way to debug this.  For now, added logging, and we'll get back to it when Aura is in better shape.

BUG=chromium-os:20583
TEST=On plain ChromeOS, can still open a file with control-O file picker.

Review URL: http://codereview.chromium.org/7941014