Get rid of synchronous calls to getting plugins to help with http://codereview.chromium.org/7980011/.
The PluginExceptionsTableModel isn't used anymore.

Change print preview and using Reader with unsupported PDFs to tell the plugin filter which plugin name they want to use, instead of giving it the plugin directly. This makes the former not need to block on getting the plugins. The later still does, to know if Reader is installed or not, however it's trivial to make PDFHasUnsupportedFeature() get the plugins asynchronously because it's just a dispatcher for an async message.
Review URL: http://codereview.chromium.org/7977042