On Windows, force the output directories to be src/build/{Debug,Release}.
This allows develoeprs to build from any sln file and not have to rebuild the
world.  This is more like how XCode projects are generated.

We choose src/build/{Debug,Release} because this is where all.sln generates
files and this is what the bots use.  This shouldn't require any changes to
the bots.

BUG=95278

Review URL: http://codereview.chromium.org/7835006