Support manually setting LKGR from script.
I needed to manually set LKGR tonight.  We used
to be able to do this from chromium-status.  Now
that the LKGR poller is shut off, the only way
to completely set it is to submit the new LKGR
to the master.

It was much simpler to just reuse the lkgr_finder
script for this with an additional argument (--manual
<LKGR_REV>).


Review URL: http://codereview.chromium.org/7970026