Implement --use-gl=any
The behavior is this: if --use-gl=any passes in, and the default GL binding causes GPU features to be blacklisted, fallback to osmesa.  Also, in GPU process, if the default binding fails, fallback to osmesa.

This option will be used in testing to make sure accelerated path is always tested, if not on GPU, then on top of osmesa.

This CL also moves certain functions out of GpuBlacklist to GpuDataManager so we could manage the special cases when --use-gl=osmesa or --use-gl=any is passed.

BUG=97675
TEST=bots green
Review URL: http://codereview.chromium.org/7967020