Correctly lookup PluginGroup for a given WebPluginInfo.
If a WebPluginInfo had an empty name, we wouldn't find it in the list of plug-in groups, and repeatedly add it, causing the list to grow every time.

Fix this by a) matching plug-ins based on the path, and b) not adding plug-ins which are not found to the PluginList.

BUG=97655
TEST=see bug


Review URL: http://codereview.chromium.org/8050014