Changed Mac, Linux and Win Official Chrome build to drop default app CRXs in the build/installer output.
This change undoes the revert of 99878.  This change was originally reverted
because it was decided to push this change to M16.

BUG=94920
TEST=Do an Official build in mac, linux and win. Ensure that after installing
the app, you can find the default apps under the install's directory
structure. For win, this is under a default_apps directory, for mac, this is
under a "Default Apps" directory.
Review URL: http://codereview.chromium.org/8041033