Add Chromium OS to the list of chromium builders.
Group the closers at the top, adding a box and a title.
Juggle the rest around to fit well.

Rework the javascript to make moving the lines around easy.  The efficiency
would not suit mobile browsers, but the ability to juggle the contents
outweighs it.

BUG=20140
TEST=run locally.

Review URL: http://codereview.chromium.org/8053001