Avoid crash when dragging text from WebUI dialogs.
This fixes a DCHECK()/segfault that happens on Chrome OS
when dragging text out of WebUI dialogs.  It currently
happens with the WebUI task manager, although we've seen
similar problems with other dialogs in the past.

r91384 added code that assumes that every TabContents has a
corresponding TabContentsWrapper, but this isn't always the
case.

BUG=chromium-os:20738,chromium-os:18781,chromium-os:17501
TEST=added a hokey unit test; also manually tested that I don't see a crash when dragging text out of the WebUI task manager

Review URL: http://codereview.chromium.org/8055033