Improve the interactivity of the stub network library.
The stub library is used when Chrome for ChromeOS is run outside of ChromeOS (e.g. on a linux desktop), with some tests, or when specified from the command line.
This CL changes the baseline stup implementation for networking, and enables actions such as connecting and disconnecting from a network.

BUG=none
TEST=All tests pass. Run chrome for chromeos on a desktop (or with --stub-cros) and test stub functionality.


Review URL: http://codereview.chromium.org/7974015