This causes the sync notifier in integration tests to broadcast notifications to all clients, including themselves.  The effect of this is that, following a commit, the client will receive notification that updates are available (even if those updates are its own) then enter a sync cycle to download them.  This means that their progress markers will be in kept in sync with the servers' latest.
For an explanation of why this is a good thing, see crbug.com/97780 and read
through the related issues.

There are several unit tests that check if the notification filtering is
working.  In order to not break those test, the notifier in the unit tests is
constructed to broadcast notifications to others only.  

BUG=97780
TEST=sync_integration_tests,sync_unit_tests


Review URL: http://codereview.chromium.org/8002031