Combine PpapiBrokerProcessHost into PpapiPluginProcessHost.
The broker does not need the network observer, so I factored that out. Since the constructors for the broker and plugin are different, I made them private and added factories that also call Init(), which must be called before the object can be used.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7979028