Disable inclusion of stats in build step status.
The UIPerfTest class implements a PrintResult() method.  One
of its arguments is a flag that determines if a given result
is "important".  The only difference in handling an important
result compared to a non-important result is that important
results are prefixed with '*' in stdio.  The rest of the
message printed to stdio is exactly the same.

This change causes most/all of these stats to no longer be
important, which means the *-prefix is removed from the
message.  By removing the *-prefix, this causes the Buildbot
to not include the result information in the step status
box in the waterfall view.

We should remove these results from the step status box since
they were originally added when there were much fewer bots
and steps so the display was not hindered by so many steps
showing many results.  And now that we auto-detect perf
regressions and improvements, the 'important' ones to watch
will be auto-included by the Buildbot.

Review URL: http://codereview.chromium.org/8139021