Allow FlushSyncs to short circuit
All messages on a channel are now executed in order, so Flushs have to wait for all previous AsyncFlushs to finish. This is generally unnecessary, so we replace Flush with a GetStateFast message. All commands are queued up when they enter the gpu process, and GetStateFast commands are moved to the head of the line.

BUG=
TEST=trybots, webgl conformance tests


Review URL: http://codereview.chromium.org/8198017