webRequest.onAuthRequired listeners can provide authentication credentials.
onAuthRequired listeners that specify "blocking" in the extraInfoSpec can return authentication credentials [username and password] in the BlockingResponse. If these are provided, Chrome will use these credentials rather than showing a login prompt for the user.

If "blocking" is not specified, or an authCredentials object is not present in the BlockingResponse, then a login prompt will be displayed.

Warning: If the authentication credentials are invalid, the extension may still present credentials for subsequent challenges. This could lead to infinite loops of bad credentials being entered without user intervention.

BUG=32056
TEST=Write an extension which does a blocking onAuthRequired and provides correct credentials, validate that it works.


Review URL: http://codereview.chromium.org/8015004