Multi-Profiles: Fix choosing random profile avatar
The code to choose the default profile avatar had a bug where it was much more likely to pick the 9th avatar icon.

I fixed the code to make it truly random. I also refactored it so that it's much easier to reason about.

BUG=99788
TEST=


Review URL: http://codereview.chromium.org/8227025