[chromiumos] Start TPM token initialization re-tries on login
The current TPM token setup logic attempts initialization once after
the user logs in. Asynchronous attempts for TPM token setup are not
triggered unless the user opens the VPN or WiFi config panel (attempt
retries are triggered via a call to CertLibraryImpl::RequestCertificates()).

This means that if the first attempt fails and the user never opens
up the WiFi config or VPN config panel, the TPM token will stay
uninitialized. This breaks the certificate manager (list of certs is
empty), the SPDY proxy extension, amongst other things.

Essentially, any part of the network subsystem that depends on the
private hardware NSS slot (via crypto::GetPrivateNSSKeySlot) stays
broken if the first attempt fails. (So, this is not just an issue with
the list of certs not being displayed correctly).

This CL changes that so that retry logic for TPM token init is
triggered right after the user logs in.

BUG=chromium-os:20933
TEST=Log in, verify from logs that TPM initialization attempts start
     immediately after. Verify that the missing certificates issue
     no longer happens.

Change-Id: I9c609bdb198a88db8ceb2019cc92c19d1983bc05


Review URL: http://codereview.chromium.org/8212003