Flash white instead of black when the URL is changed in the omnibox for the same tab
This is separate from the previous fix, when a new tab is opened.
When the URL of a tab is changed, the RWHV is deleted and another one is recreated. There is a period of time for which the root view contains no RWHV and is the top layer

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/8208009