[Sync] Remove unnecessary DCHECK in pref model associator.
We assumed that when a preference becomes unmanaged, it will be user
controlled as well. This isn't the case if it was a default value previously.
This gets handled properly in the association code though, so just remove the
DCHECK and make the comment clearer.

BUG=97611
TEST=unit_tests --gtest_filter="*ProfileSyncServicePreference*"


Review URL: http://codereview.chromium.org/8266001