Support SSL connections in websocket-to-TCP proxy.
For historical reasons current implementation of WS-to-TCP proxy was implemented as standalone libevent-based server.
Then it was integrated into chromium as is.
In order to support SSL we need to connect libevent-based proxy with MessageLoopForIO-based chromium network stack.
We do it using pipes.

It is intended as temporary solution until we will have
new shiny implementation of WS-to-TCP proxy integrated into network stack.

BUG=chromium-os:15533
TEST=Manual

Review URL: http://codereview.chromium.org/8087001