Downgrade a CHECK to a NOTREACHED in ResourceBundle::LoadLocaleResources.
This is a case where we're failing to mmap the file, even though it exists
on disk.  For example, we can't read the file or it has a size of 0.

We don't need to CHECK because in chrome_browser_main.cc, we prompt the
user to reinstall if this happens.

BUG=100395


Review URL: http://codereview.chromium.org/8301017