chromeos: Implement the new D-Bus client for SessionManagerClient.
This is part 1 of the LoginLibrary to SessionManagerClient migration.
The new files are added but not used from Chrome yet.

The new code is based on login_library.cc and chromeos_login.cc,
and the new code will replace the them.

For code review, please refer to the files below:

http://src.chromium.org/viewvc/chrome/trunk/src/chrome/browser/chromeos/cros/login_library.h?view=markup
http://src.chromium.org/viewvc/chrome/trunk/src/chrome/browser/chromeos/cros/login_library.cc?view=markup
http://git.chromium.org/gitweb/?p=chromiumos/platform/cros.git;a=blob;f=chromeos_login.h;hb=HEAD
http://git.chromium.org/gitweb/?p=chromiumos/platform/cros.git;a=blob;f=chromeos_login.cc;hb=HEAD

Per suggestion from cmasone, we'll be using SessionManagerBlah rather
than LoginManagerBlah for new code, as the former is more appropriate.

BUG=chromium-os:16555
TEST=chrome and tests build

Review URL: http://codereview.chromium.org/8295019